requirements.txt
streamlit==1.29.0
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
keras==2.11.0
